Add full client side validation to signup form. Move influencer create to API endpoint.
