+++ app/routes/api/index.js
@@ -3,9 +3,11 @@ const router = express.Router();
 const adminRoutes = require('./admin');
 const campaignRoutes = require('./campaign');
 const eventRoutes = require('./event');
+const influencerRoutes = require('./influencer');
 
 router.use('/', adminRoutes);
 router.use('/', campaignRoutes);
 router.use('/', eventRoutes);
+router.use('/', influencerRoutes);
 
 module.exports = router;
