+++ app/routes/api/influencer.js
@@ -1,0 +1,72 @@
+const express = require('express');
+const router = express.Router();
+var passport = require('passport');
+const Influencer = require('../../models/influencer');
+const Signup = require('../../models/signup');
+
+router.route('/influencers')
+
+.post(function(request, response) {
+    var findStr = { email: request.body.email, signupLink: request.body.signupLink };
+    Signup.findOne(findStr).exec()
+    .then(function(signup) {
+        if (signup) {
+            Influencer.register(
+                new Influencer({
+                    firstName: request.body.firstName,
+                    lastName: request.body.lastName,
+                    username: request.body.username,
+                    email: request.body.email
+                }),
+                request.body.password,
+                function(err, account) {
+                    if (err) {
+                        console.log('error! could not create new influencer: ' + err);
+                        response.status(403).send({
+                            error: {
+                                statusCode: 403,
+                                status: 'Unauthorized',
+                                message: 'There was an error creating account.'
+                            }
+                        });
+                    } else {
+                        Signup.remove(findStr, function() {
+                            passport.authenticate('local')(request, response, function() {
+                                response.status(200).end();
+                            }, function() {
+                                response.status(403).send({
+                                    error: {
+                                        statusCode: 403,
+                                        status: 'Unauthorized',
+                                        message: 'There was an error logging in.'
+                                    }
+                                });
+                            });
+                        });
+                    }
+                }
+            );
+        } else {
+            console.log('Influencer not found in signup database! Influencer: ' + request.body.email + ' link: ' + request.body.signupLink);
+            response.status(403).send({
+                error: {
+                    statusCode: 403,
+                    status: 'Unauthorized',
+                    message: 'Influencer not found in signup database.'
+                }
+            });
+        }
+    })
+    .catch(function(reason) {
+        console.log('Error! Could not find signup link: ' + reason);
+        response.status(403).send({
+            error: {
+                statusCode: 403,
+                status: 'Unauthorized',
+                message: 'Influencer not found in signup database.'
+            }
+        });
+    });
+});
+
+module.exports = router;
