+++ app/views/signup.hbs
@@ -7,29 +7,45 @@
             <div class="col-sm-7 col-sm-offset-3">
                 <form name="signupForm">
                     <div class="form-group">
-                        <input type="text" class="form-control input-lg" placeholder="Enter your signup email here (required)" ng-model="signupDetails.email"
-                            ng-required="true">
-                    </div>
-                    </br>
-                    </br>
-                    <div class="form-group">
-                        <input type="text" class="form-control input-lg" placeholder="First name" ng-model="signupDetails.firstName" ng-required="true">
+                        <input name="firstName" type="text" class="form-control input-lg" placeholder="First name"
+                            ng-model="signupDetails.firstName" ng-required="true"
+                            ng-class="{'form-field-error': signupForm.firstName.$touched && signupForm.firstName.$invalid}">
                     </div>
                     <div class="form-group">
-                        <input type="text" class="form-control input-lg" placeholder="Last name" ng-model="signupDetails.lastName" ng-required="true">
+                        <input name="lastName" type="text" class="form-control input-lg" placeholder="Last name"
+                            ng-model="signupDetails.lastName" ng-required="true"
+                            ng-class="{'form-field-error': signupForm.lastName.$touched && signupForm.lastName.$invalid}">
                     </div>
                     <div class="form-group">
-                        <input type="text" class="form-control input-lg" placeholder="Username" ng-model="signupDetails.username" ng-required="true">
+                        <input name="email" type="email" class="form-control input-lg" placeholder="Email address"
+                            ng-model="signupDetails.email" ng-required="true"
+                            ng-class="{'form-field-error': signupForm.email.$touched && signupForm.email.$invalid}">
+                        <span class="error-help-text" ng-if="signupForm.email.$touched && signupForm.email.$error.email">Please enter a valid email address.</span>
                     </div>
                     <div class="form-group">
-                        <input type="password" class="form-control input-lg" placeholder="Password" ng-model="signupDetails.password" ng-required="true">
+                        <input name="username" type="text" class="form-control input-lg" placeholder="Username"
+                            autocomplete="off" autocorrect="off" autocapitalize="off"
+                            ng-model="signupDetails.username" ng-required="true"
+                            ng-class="{'form-field-error': signupForm.username.$touched && signupForm.username.$invalid}">
                     </div>
                     <div class="form-group">
-                        <input type="password" class="form-control input-lg" placeholder="Retype password" ng-model="signupDetails.passwordConfirm"
-                            ng-required="true" match-input="signupDetails.password">
+                        <input name="password" type="password" class="form-control input-lg" placeholder="Password"
+                            ng-model="signupDetails.password" ng-required="true"
+                            ng-class="{'form-field-error': signupForm.password.$touched && signupForm.password.$invalid}">
+                    </div>
+                    <div class="form-group">
+                        <input name="passwordConfirm" type="password" class="form-control input-lg" placeholder="Retype password"
+                            ng-model="signupDetails.passwordConfirm" ng-required="true" match-input="signupDetails.password"
+                            ng-class="{'form-field-error': signupForm.passwordConfirm.$touched && signupForm.passwordConfirm.$invalid}">
+                        <span class="error-help-text" ng-if="signupForm.passwordConfirm.$touched && signupForm.passwordConfirm.$error.unique">Passwords must match.</span>
                     </div>
                     <input type="hidden" ng-init="signupDetails.signupLink='{{{signup.signupLink}}}'">
-                    <button type="submit" ng-click="signUp()" class="btn primary-button" ng-disabled="signupForm.$invalid">Sign Up</button>
+                    <div class="errors" ng-if="error">
+                        {[{error}]}
+                    </div>
+                    <div class="form-group">
+                        <button type="submit" ng-click="signUp()" class="btn primary-button" ng-disabled="signupForm.$invalid">Sign Up</button>
+                    </div>
                 </form>
             </div>
