+++ public/js/index.js
@@ -8,4 +8,3 @@ require('./signup.controller');
 require('./success.controller');
 require('./user.controller');
 require('./visitor.controller');
-
