+++ public/scss/_colors.scss
@@ -2,3 +2,4 @@ $yellow: #FDCC06;
 $blue-light: #35A5B3;
 $white: #FFF;
 $gray-light: #e5e5e5;
+$dark-red: #ad2a2a;
